Splitting options into separate file
