+++ .gitignore
@@ -1,3 +1,5 @@
+.idea/
+
 out.txt
 out*.txt
 old*.txt
